requirements.txt
pyzmq
ujson
